@@ -2,8 +2,13 @@ import * as React from 'react'
 
 import ResumeMetadata from 'data/ResumeMetadata'
 
-export default (props: any) => {
+type TProps = React.HTMLAttributes<HTMLDivElement>
+export default (props: TProps) => {
+  const {
+    objective,
+  } = ResumeMetadata
+  const ps = objective.split('\n').map(p => <p key={p}>{p}</p>)
   return (
-    <span {...props}>{ResumeMetadata.objective}</span>
+    <span {...props}>{ps}</span>
   )
 }
